Use ninja-build to reduce build times; fix build on Linux

+++ iris/wsi/platform_window_x11.h
@@ -51,10 +51,12 @@ public:
   Offset2D Offset() const noexcept { return rect_.offset; }
   Extent2D Extent() const noexcept { return rect_.extent; }
 
-  /*! \brief Get the current cursor position in screen coordinates.
-   *  \return the current cursor position in screen coordinates.
+  /*! \brief Get the current cursor position in screen coordinates if this
+   * window is the active window.
+   * \return the current cursor position in screen coordinates if this window is
+   * the active window, or (-FLT_MAX, -FLT_MAX)
    */
-  glm::uvec2 CursorPos() const noexcept;
+  glm::vec2 CursorPos() const noexcept;
 
   std::string Title() noexcept {
     //::XTextProperty prop;
